refactor: 참가자 생성 책임 분리
