+++ src/main/java/blackjack/BlackjackRunner.java
@@ -1,46 +1,38 @@
 package blackjack;
 
-import blackjack.domain.Amount;
 import blackjack.domain.Deck;
 import blackjack.domain.Hand;
-import blackjack.domain.Nicknames;
 import blackjack.domain.Participants;
-import blackjack.domain.Players;
-import blackjack.domain.participant.Dealer;
 import blackjack.dto.DrawResult;
 import blackjack.dto.ParticipantResult;
 import blackjack.dto.PlayerNicknamesResult;
 import blackjack.dto.TotalGameResult;
-import blackjack.util.PlayerNameParser;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 import java.util.List;
 
 public class BlackjackRunner {
 
-    private final InputView inputView;
+    private final ParticipantsCreator participantsCreator;
     private final OutputView outputView;
+    private final InputView inputView;
 
     public BlackjackRunner(InputView inputView, OutputView outputView) {
         this.inputView = inputView;
         this.outputView = outputView;
+        this.participantsCreator = new ParticipantsCreator(inputView, outputView);
     }
 
     public void execute() {
-        Participants participants = makeParticipants();
+        Participants participants = participantsCreator.create();
         Deck deck = Deck.createShuffledDeck();
-        deck = gameStart(participants, deck);
+        deck = participants.distributeCards(deck);
 
         printInitialSetup(participants);
         printInitialResult(participants);
-
         deck = playerTurn(participants, deck);
         dealerTurn(participants, deck);
-        gameEnd(participants);
-    }
-
-    private Deck gameStart(Participants participants, Deck deck) {
-        return participants.distributeCards(deck);
+        printGameResult(participants);
     }
 
     private void printInitialSetup(Participants participants) {
@@ -48,57 +40,9 @@ public class BlackjackRunner {
         outputView.printInitialSetUp(playerNicknamesResult);
     }
 
-    private void gameEnd(Participants participants) {
-        printGameResult(participants);
-    }
-
     private void printInitialResult(Participants participants) {
         List<ParticipantResult> participantResults = participants.getInitialResult();
         outputView.printInitialResult(participantResults);
-    }
-
-    private Participants makeParticipants() {
-        Nicknames nicknames = getPlayersNickname();
-        Players players = betPlayers(nicknames);
-        Dealer dealer = Dealer.from();
-        return new Participants(players, dealer);
-    }
-
-    private Nicknames getPlayersNickname() {
-        try {
-            outputView.askGameMembers();
-            String playerNamesInput = inputView.readLine();
-            List<String> playerNames = PlayerNameParser.parsePlayerNames(playerNamesInput);
-            return Nicknames.from(playerNames);
-        } catch (IllegalArgumentException e) {
-            outputView.printLine(e.getMessage());
-            return getPlayersNickname();
-        }
-    }
-
-    private Players betPlayers(Nicknames nicknames) {
-        List<String> playerNicknames = nicknames.getRawNicknames();
-        Players players = Players.makeEmptyPlayers();
-        for (String nickname : playerNicknames) {
-            players = addPlayerWithValidBet(nicknames, nickname, players);
-        }
-        return players;
-    }
-
-    private Players addPlayerWithValidBet(Nicknames nicknames, String nickname, Players players) {
-        try {
-            Amount playerRequest = readBettingRequest(nickname);
-            return players.addPlayer(nicknames.findByNickname(nickname), playerRequest);
-        } catch (IllegalArgumentException e) {
-            outputView.printLine(e.getMessage());
-            return addPlayerWithValidBet(nicknames, nickname, players);
-        }
-    }
-
-    private Amount readBettingRequest(String nickname) {
-        outputView.askBetAmount(nickname);
-        String amount = inputView.readLine();
-        return Amount.from(amount);
     }
 
     public void printGameResult(Participants participants) {
@@ -147,6 +91,6 @@ public class BlackjackRunner {
         if (!(isDraw || isNotDraw)) {
             throw new IllegalArgumentException("유효한 커맨드(예는 y, 아니오는 n)를 입력해 주세요.");
         }
-        return userCommand.equals("y");
+        return isDraw;
     }
 }
